backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic[email]>=2.0
email-validator>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
motor>=3.0
pymongo>=4.0
passlib[bcrypt]>=1.7
bcrypt<4.1
python-jose[cryptography]>=3.3
openai>=1.0
Pillow>=10.0
